[AL-10018] Split async-sync client and balancer
