+++ sdc_client/__init__.py
@@ -1,4 +1,6 @@
 from .interfaces import ILogger, IStreamSets, IStreamSetsProvider, IPipelineProvider, IPipeline
 from .client import *
+from .async_client import *
 from .balancer import StreamsetsBalancer
+from .async_balancer import StreamsetsBalancerAsync
 from .api_client import ApiClientException, UnauthorizedException
